+++ columns/columns_model.js
@@ -1,9 +1,9 @@
-const db = require("../database/db-config");
+// const db = require("../database/db-config");
 
-function findColumn() {
-  return db("columns").select();
-}
+// function findColumn() {
+//   return db("columns").select();
+// }
 
-module.exports = {
-  findColumn,
-};
+// module.exports = {
+//   findColumn,
+// };
