+++ tags/tags_router.js
@@ -25,11 +25,11 @@ router.post("/addTag/:id", async (req, res, next) => {
     const newTag = await tagMod.addTag(req.body, id);
     if (newTag) {
       res.status(201).json({
-        message: "New Tag Created"
+        message: "New tag created"
       });
     } else {
       send.status(500).json({
-        message: "Error Saving New Tag, please try again later"
+        message: "Error saving new tag, please try again later"
       });
     }
   } catch (err) {
